work on pdf tests.

+++ geocat-services/src/main/java/org/fao/geonet/geocat/TestPdfExportAllMd.java
@@ -39,20 +39,23 @@ public class TestPdfExportAllMd {
         long lastFlushResults = lastPrintProgress;
         int i = 0;
         for (Integer id : ids) {
-            i++;
-            if (System.currentTimeMillis() - lastPrintProgress > 30000) {
-                System.out.println("\n\n==============\n" + i + " / " + ids.size() + " \n===================\n");
-                lastPrintProgress = System.currentTimeMillis();
-            }
-            if (System.currentTimeMillis() - lastFlushResults > (60000 * 10)) {
-                Files.write(Paths.get("/tmp/PDFExportReport.txt"), response.toString().getBytes());
-                lastFlushResults = System.currentTimeMillis();
-            }
-            for (String l : new String[]{"eng", "ger", "fra", "ita"}) {
-                try {
-                    format.exec(l, "pdf", id.toString(), "full_view", "y", true, FormatterWidth._100, new ServletWebRequest(new MockHttpServletRequest(), new MockHttpServletResponse()));
-                } catch (Throwable t) {
-                    response.append("\n - ").append(l).append(": ").append(id);
+            if (repo.exists(id)) {
+                i++;
+                if (System.currentTimeMillis() - lastPrintProgress > 30000) {
+                    System.out.println("\n\n==============\n" + i + " / " + ids.size() + " \n===================\n");
+                    lastPrintProgress = System.currentTimeMillis();
+                }
+                if (System.currentTimeMillis() - lastFlushResults > (60000 * 10)) {
+                    Files.write(Paths.get("/tmp/PDFExportReport.txt"), response.toString().getBytes());
+                    lastFlushResults = System.currentTimeMillis();
+                }
+                for (String l : new String[]{"eng", "ger", "fra", "ita"}) {
+                    try {
+                        format.exec(l, "pdf", id.toString(), "full_view", "y", true, FormatterWidth._100, new ServletWebRequest(new MockHttpServletRequest(), new MockHttpServletResponse()));
+
+                    } catch (Throwable t) {
+                        response.append("\n - ").append(l).append(": ").append(id).append(" - ").append(t.getMessage());
+                    }
                 }
             }
         }
